Kludge the meta description code some more (strip <a> tags whose inner text begins with 'http://')
Many users start threads with a link. Not many take care to attach a
textual description to it. As a result, we waste many characters of our
meta descriptions to store a URL which is probably of very little use to
those making use of the description itself.

This commit largely mitigates this problem by getting rid of all links
whose inline text starts with 'http://'. It's not even trying to be a
catch-all - there could be false negatives and false positives. A better
solution may exist, but I can't think of one. Considering the relatively
low importance of meta descriptions, I'd rather fix most of them and
break a tiny subset than have the larger subset broken in the first
place.