Limit the maximum width of images through JavaScript
This commit introduces the following changes:

1. Modify the behaviour of the JavaScript image in following ways:
a) It will now be applied to all images of class 'bbc_img', previously
'bbc_img resized'
b) A max-width of 700px is imposed on top of the user-defined width (if
present)

2. Themes/tintagel_fes/scripts/theme.js has been removed and the
reference to that file in the theme code has been replaced with the same
file from the default theme. Prior to the changes in point 1, the two
files were identical.