Add "Top Lurkers" to forum stats
Generate a list of top 10 lurkers based on time_online/num_posts. It
could be a good idea to set no_posts to something higher than 0 to
prevent people with just a handful of posts from dominating the stats.
Then again, if they lurked for hours and only posted once, then maybe
they deserve it?