Allow people to unsubscribe from threads they've posted in
This patch adds an "unsubscribe" button to each thread on the unread
replies page, as well as to the thread view itself. This will only
appear if the user has posted in the thread (as that is the only way it
will appear on the unread replies page), and if the user unsubscribes,
it turns into a "subscribe" button.

Currently, posting in an unsubscribed thread does *not* cause automatic
re-subscription. This is being discussed in the below thread, and may
be added later:

  http://forum.tfes.org/index.php?topic=2550.msg65143#msg65143

This patch also re-orders the Christmas Season 2 buttons to be
consistent with tintagel_fes, and to be the same at the top and bottom
of a thread.