smc_Toggle: Allow specifying arbitrary triggers
Previously, smc_Toggle had two possible ways of registering a trigger
for toggle events. Either an image may be specified in aSwapImages, or
a link may be specified in aSwapLinks. In either case, the semantics
are overloaded; passing such an object means *both* that it is a
trigger for the toggle and that it is to be toggled, with no way to get
one without the other.

In the blanko_fes theme, we don't use an image as our toggle trigger,
and our "upshrink" button doesn't need to be toggled. Until now, we've
been pretending it's an image so that smc_Toggle will happily respond
to click events from it, but this had a couple of side effects:

 1. On initial page load, the URL "/undefined" would be preloaded in
    anticipation of toggling an image to some other state, except we
    hadn't provided a URL to toggle to because there is no actual
    image. This is a spurious HTTP request that clutters logs at best
    and hampers performance at worst.

 2. When clicked, the "upshrink" button would have its title and alt
    text set to "undefined", because we hadn't provided these either.
    This has rather poignant implications for accessibility.

Instead, we now add a new aSwapTriggers key that triggers the toggle on
a click event without trying to do anything else. That's all we need,
and now it's all we'll get.

Also, since smc_Toggle no longer unhides our toggle button for us
thinking it's an image, don't set "display: none" in its definition.