Fix caching where $boardurl is not static
Previously, $boardurl was treated as static configuration by some parts
of the code, and as mutable data by others. In particular, the caching
implementation uses $boardurl to generate cache keys, which results in
two or more different caches being accessed at different times during
request processing if $boardurl gets modified.

There are three situations in which $boardurl may be modified, all of
them in loadTheme in Load.php:

 1. $modSettings["forum_alias_urls"] is set, and the URL being accessed
    matches one of its entries.
 2. The URL being accessed differs only by scheme, using https instead
    of http.
 3. The site is being accessed by IP address rather than DNS name.

Any of these cases will trigger this bug.

There are numerous potential symptoms of reading from a stale cache,
but the most obvious one we noticed was that boards containing recent
posts made over HTTPS continue to be marked as unread, even once those
posts are read. This is due to the way unread board status is tracked.

SMF tracks the most recent post ID on the entire site in
$modSettings["maxMsgID"], which is read from cache very early in
request processing. When a user marks a board as read by viewing its
index, the log_boards table in the database is updated to indicate that
their most recent read post is $modSettings["maxMsgID"] -- in other
words, any posts made after now will be considered unread.

However, when a user makes a post over HTTPS, $modSettings["maxMsgID"]
gets updated and written to both the database and the *wrong* cache.
Next time a page is loaded (over HTTP or HTTPS; it doesn't matter
because loading $modSettings from cache happens before modifying
$boardurl), it gets set back to its old value, and marking a board as
read then sets one's most recent read post to one less than it should
be. That board is thus considered to still contain an unread post.

The root of the problem, as previously mentioned, is that some parts of
the code treat $boardurl as static configuration, and others as mutable
data. The fix here is simply to separate these use cases; store the
original value as $canonical_boardurl, and use that wherever a
consistent identifier for the board is required.