Make quick edit a toggleable option
Putting the quick edit button in a menu that already has a modify
button is somewhat self-defeating -- the whole point of quick edit is
that it's right there and you can do it quickly.

Since apparently not everyone wants to see it, allow users to toggle
its visibility in their preferences.