Don't use "hidden" class in Who's Online
The "hidden" class isn't styled anywhere in the CSS in any of our
themes, but it *is* styled in Bootstrap's CSS as used in blanko_fes.
Surprise surprise, putting something in the "hidden" class hides it
with Bootstrap. This was a terrible choice of class name (on SMF's
part).

Since this does the wrong thing in one theme and does nothing in any of
the others, just get rid of it.