Set $utf8 to true
It appears that this variable's only purpose is to ensure that we are not running PHP < 4.2.3 on Windows. This is an extremely unlikely scenario these days, and other parts of this check appear to be broken (namely, we're checking the contents of $txt and $modSettings before they're even set).

$utf8 is checked in a fair amount of places and guides many of the application's assumptions - the issue was originally found while investigating edge cases in $smcFunc['substr']

It is still a good idea to investigate this further - clearly someone assumed that these variables would have been set by this point, and the fact that they aren't could be a symptom of a larger issue