Use imagemagick to resize images
SMF's default behaviour for image resizing is to use GD (and to do so badly). This commit replaces the `resizeImageFile()` function with one relying on ImageMagick

`checkImagick()` was also introduced to verify if the `Imagick` class is available. If not, `resizeImageFile()` will fall back on the old function (now named `legacyResizeImageFile()`.

The original `resizeImageFile()` code includes some code to download the file if the source filepath starts with "http://". This has been altered to "http://" or "https://", but this appears to be unused at first glance, and should be considered for deletion.

Several functions which run `resizeImageFile()` run `checkGD()` first to see if GD is available. These need to be reviewed to see if this behaviour is still desirable.