Stop mangling UTF-8 when quick-editing posts
SMF's JS is somehow worse than their PHP. The way strings are encoded, re-encoded and re-re-un-re-encoded is inconsistent between use cases. This commit removes two arbitrary invokations of `escape()` which appear to mangle many UTF-8 characters with no visible benefit. There are still a couple of undocumented and unexplained regex substitutions, but I chose not to touch them - trying to fix all of this is a rabbit hole I'm not prepared to delve into just yet.