Do not draw a completely empty span next to posts
The original theme used this span to draw a little triangle next to the post, making it look like a little comic book speech bubble. We don't do that, and this <span> was just left in with display:none. Instead of making it invisible, let's just not include it in the HTML.