Do not create empty <div>s for last modified
There really is no reason to generate these for every post, since most of them are never going to be used.

topic.js relies on the existence of these, because it wants to overzealously edit them if a post has been edited. I don't think this functionality is worth keeping, so I'm just getting rid of it