Don't generate next/prev links when impossible
If a user tries to look at a thread, but is unable to (e.g. due to lacking privileges), the forum still attempts to generate a set of rel next/prev links. This can't work, because we have no access to any details of the thread involved. Shockingly, this results in error log spam.

Unfortunately, SMF offers no variable for "a thread has been selected and is also actually being viewed", so fixing this is a bit hacky. I opted to assume that any thread with pages that can be viewed will have a page index.

These next/prev links are bordering on deprecated, so risk of regression is minimal. We *could* consider simply removing these links altogether, but technically they carry some semantic value, and for all we know there might be one piece of software somewhere that actually uses them.