Don't generate a mark read button in recent posts
The ["Recent Posts" view](https://forum.tfes.org/index.php?action=recent) of blanko_fes currently attempts to generate a "mark all posts as read" button by invoking an unitialised variable. This obviously fails, producing no output to users other than logging a few errors in our precious log.

I can think of two approaches of fixing this: either actually define the button, or stop trying to generate it. My logic is that since the button was never there for users, there is no need to add it now. On top of that, a button to mark all posts as read could be detrimental in a context which does not clearly illustrate which posts would and wouldn't be affected.