Separate hostname bans from IPv4 insanity
Originally, SMF would only check for hostname bans if the user's IP address passed the gauntlet of insanity (lots of convoluted code that largely attempts to check if it's a valid IPv4 address).

This check appears to be unnecessary - if for whatever reason the IP address is malformed, `host_from_ip()` will simply return `false`, and the rest of the hostname ban check will be skipped. In addition, original behaviour prevented hostname checks to ever happen if the user's IP is IPv6.

This commit moves the hostname check outside of the IP check (though the IP check needs a thorough rewrite further down the line).